Fix priority not applying.
Note, this seems to be a greedier approach to handling rule priority as we create a decoration type per update. But, this seems to be the way the internal vs code logic functions off of. I noticed a slight performance degradation after this, but if we are looking for a culprit in the future, we can remove the decoration type generation per decoration apply.

This commit also causes a drop event to be sent regardless of location. It seems this was the intended logic on the extension side but was never propagated to the webviews.